+++ Assets/Scripts/Enemy/AI/EnemyAI.cs
@@ -1,0 +1,11 @@
+﻿using UnityEngine;
+
+namespace Enemy.AI
+{
+  public abstract class EnemyAI : MonoBehaviour
+  {
+    public abstract void StartAI();
+    
+    
+  }
+}
