+++ Assets/Scripts/Enemy/AI/ShieldAI.cs.meta
@@ -1,0 +1,3 @@
+﻿fileFormatVersion: 2
+guid: 04e808288ccc42449872489521bfa863
+timeCreated: 1693973772
